pyproject.toml
[project]
name = "jobs-watcher"
version = "0.1.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.20.0.post0",
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-migrate>=4.0.0",
    "psycopg2-binary>=2.9.0",
    "sentry-sdk[flask]>=2.29.1",
    "python-jobspy>=1.1.80",
]
